3.3-permissions/api_with_restrictions/requirements.txt
django~=5.1
djangorestframework~=3.15.2
psycopg2-binary
django-filter~=24.3

python-dotenv~=1.0.1